Probable usable data so far
I'm using Foursquare data for some 500 venues locations around a table of 18 localities in Bogota, Youll find the localities mappend and even an attempt to clustering based on the venues, but it seems to not be working as well, cause I'm clustering localities based on venue distribution. Nontheless, its a decent starter so that I can eventually map and cluster venues and predict a new venue's amount of digital interaction based on proximity to most digitally relevant parts of the city.